+++ StorageScreen/src/main/java/com/xxmrk888ytxx/storagescreen/AudioStorageList/contracts/ProvideFileByAudioId.kt
@@ -12,5 +12,5 @@ import java.io.File
  */
 interface ProvideFileByAudioId {
 
-    suspend fun provide(id:Int) : File?
+    suspend fun provide(id: Long) : File?
 }
